WIP: removed unnecessary code, connection error handling
