+++ migrafana/src/core/api/base.py
@@ -1,27 +1,29 @@
 from grafana_client import GrafanaApi
-from core.models import GrafanaConfig
+
+from api.models import GrafanaConnection, GrafanaConfig, GrafanaConnectionError
+
+def connect(conf: GrafanaConfig):
+    grafana_inst = GrafanaApi.from_url(
+        url=conf.url,
+        credential=(conf.creds.login, conf.creds.password)  # Tuple of (user, pass)
+    )
+    try:
+        grafana_inst.connect()
+        return GrafanaConnection(instance=grafana_inst)
+    except Exception as e:
+        return GrafanaConnection(error=e)
 
 
-class GrafanaAPIClient:
-    def __init__(self, config: GrafanaConfig):
-        self.config = config
-        self.client = self._initialize_client()
+class GrafanaBaseManager():
 
-    def _initialize_client(self) -> GrafanaApi:
-        if self.config.api_key:
-            return GrafanaApi.from_url(
-                url=self.config.url,
-                credential=self.config.api_key
-            )
-        elif self.config.username and self.config.password:
-            return GrafanaApi.from_url(
-                url=self.config.url,
-                credential=(self.config.username, self.config.password)
-            )
-        raise ValueError("No valid authentication method provided")
+    def __init__(self):
+        self.connection = None
 
-    def test_connection(self) -> bool:
-        try:
-            return bool(self.client.connect())
-        except Exception:
-            return False
+    @classmethod
+    def new(cls, conf: GrafanaConfig) -> GrafanaConnection:
+        connection: GrafanaConnection = connect(conf)
+        if connection.error:
+            logger.error()
+            raise GrafanaConnectionError
+
+
