+++ migrafana/src/core/api/models.py
@@ -3,13 +3,23 @@ from pydantic import BaseModel, ConfigDict
 from grafana_client import GrafanaApi
 
 
-class NoCredsError(BaseException):
+class NoCredsError(Exception):
     ...
 
 
-class GrafanaCreds(BaseModel):
+class GrafanaConnectionError(Exception):
+    ...
+
+
+class Creds(BaseModel):
     login: str
     password: str
+    token: str
+
+
+class GrafanaConfig(BaseModel):
+    url: str
+    creds: Creds
 
 
 class GrafanaConnection(BaseModel):
